requirements.txt
alabaster==0.7.12
aniso8601==4.0.1
asn1crypto==0.24.0
Babel==2.6.0
bleach==3.0.2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
CommonMark==0.5.4
coverage==4.5.3
cryptography==2.4.2
docutils==0.14
enum34==1.1.6
filelock==3.0.10
Flask==1.0.2
Flask-BasicAuth==0.2.0
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
gitdb2==2.0.3
gunicorn==19.9.0
heroku==0.1.4
idna==2.7
imagesize==1.1.0
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mongomock==3.14.0
mqfactory==0.0.4
packaging==18.0
paho-mqtt==1.4.0
persist-queue==0.4.1
pkginfo==1.4.2
pluggy==0.8.0
psutil==5.4.8
py==1.7.0
py-bcrypt==0.4
pycparser==2.19
Pygments==2.2.0
pymongo==3.7.2
pyparsing==2.3.0
python-dateutil==2.7.2
python-dotenv==0.8.2
pytz==2018.7
readme-renderer==24.0
recommonmark==0.4.0
requests==2.20.1
requests-toolbelt==0.8.0
sentinels==1.0.0
servicefactory==0.9
simple-rsa==1.0.0
six==1.11.0
smmap2==2.0.5
snowballstemmer==1.2.1
Sphinx==1.7.9
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
toml==0.10.0
tox==3.6.1
tqdm==4.28.1
twine==1.12.1
typing==3.6.6
urllib3==1.24.1
virtualenv==16.1.0
webencodings==0.5.1
Werkzeug==0.14.1
